Refactor configuration system
move to separate module and update imports
