+++ modules/annoyance.mjs
@@ -1,16 +1,7 @@
 import { EmbedBuilder, Events } from 'discord.js';
-import config from '../config.json' with { type: 'json' };
 import { BotModule } from './util/module.mjs';
 
 export default class AnnoyanceModule extends BotModule {
-	/**
-	 * Cache color configuration here + TS assertions
-	 * @constant
-	 */
-	colors = {
-		RED: /** @type {`#${string}`} */ (config.red),
-		GREEN: /** @type {`#${string}`} */ (config.green)
-	};
 	constructor () {
 		super('Annoyance', 'sus module sus module', ['messageCreate']);
 	}
